requirements.txt
Flask==3.0.0
firebase-admin
python-dotenv
Werkzeug
email-validator
